Projects/cheongju/cheongju_llm_rag/requirements.txt
streamlit==1.41.1
pandas==2.2.3
numpy==2.1.3
python-dotenv==1.0.1
openai==1.57.4
faiss-cpu==1.9.0
tqdm==4.67.1
